Product's CRUD fixes

+++ src/controllers/product.ts
@@ -199,30 +199,35 @@ export const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { id } = req.params
 
-    // 1. check if it has already been ordered
+    const ref = fb.doc(collections.products, id)
+    const product = await fb.getDoc(ref)
 
-    const query = fb.query(
-      collections.orders,
-      fb.where("productsIds", "array-contains", id)
-    )
-    const docsSnap = await fb.getDocs(query)
+    if (product.exists()) {
+      // 1. check if it has already been ordered
 
-    if (docsSnap.docs.length === 0) {
-      // 2. if not, delete
-
-      const ref = fb.doc(collections.products, id)
-      await fb.deleteDoc(ref)
-
-      res.status(200).json({ success: true })
-    } else {
-      // 3. if true, it cant be deleted. Just inactivated
-      throw new Error(
-        JSON.stringify({
-          message:
-            "Este produto já foi pedido pelo menos uma vez. Em vez de deletar, inative-o.",
-        })
+      const query = fb.query(
+        collections.orders,
+        fb.where("productsIds", "array-contains", id)
       )
-    }
+      const docsSnap = await fb.getDocs(query)
+
+      if (docsSnap.docs.length === 0) {
+        // 2. if not, delete
+
+        const ref = fb.doc(collections.products, id)
+        await fb.deleteDoc(ref)
+
+        res.status(200).json({ success: true })
+      } else {
+        // 3. if true, it cant be deleted. Just inactivated
+        throw new Error(
+          JSON.stringify({
+            message:
+              "Este produto já foi pedido pelo menos uma vez. Em vez de deletar, inative-o.",
+          })
+        )
+      }
+    } else throw new Error("Este produto não existe.")
   } catch (error) {
     res.status(400).json(getCustomError(error))
   }
@@ -230,14 +235,14 @@ export const deleteProduct = async (req: Request, res: Response) => {
 
 export const toggleProductStatus = async (req: Request, res: Response) => {
   try {
-    const data = req.body
+    const { id } = req.params
 
-    const validation = productValidator(data)
+    const ref = fb.doc(collections.products, id)
+    const product = await fb.getDoc(ref)
 
-    if (validation.ok) {
-      const ref = fb.doc(collections.products, data.id)
-      await fb.updateDoc(ref, { active: !data.active })
-      const docData = data
+    if (product.exists()) {
+      await fb.updateDoc(ref, { active: !product.data().active })
+      const docData = { ...product, active: !product.data().active }
 
       res.status(200).json({ success: true, data: docData })
     } else {
@@ -245,6 +250,8 @@ export const toggleProductStatus = async (req: Request, res: Response) => {
         "Não foi possível alterar o status. Tente novamente mais tarde."
       )
     }
+
+    const data = req.body
   } catch (error) {
     res.status(400).json(getCustomError(error))
   }
